Refactor extensions into separate target
