+++ Sources/CSQLite/csqlite_shims.c
@@ -245,71 +245,3 @@ int csqlite_sqlite3_vtab_config_directonly(sqlite3 *db)
 {
 	return sqlite3_vtab_config(db, SQLITE_VTAB_DIRECTONLY);
 }
-
-// MARK: - Database extensions
-
-void sqlite3_carray_init();
-int csqlite_sqlite3_auto_extension_carray()
-{
-	return sqlite3_auto_extension(sqlite3_carray_init);
-}
-
-int csqlite_sqlite3_cancel_auto_extension_carray()
-{
-	return sqlite3_cancel_auto_extension(sqlite3_carray_init);
-}
-
-void sqlite3_decimal_init();
-int csqlite_sqlite3_auto_extension_decimal()
-{
-	return sqlite3_auto_extension(sqlite3_decimal_init);
-}
-
-int csqlite_sqlite3_cancel_auto_extension_decimal()
-{
-	return sqlite3_cancel_auto_extension(sqlite3_decimal_init);
-}
-
-void sqlite3_ieee_init();
-int csqlite_sqlite3_auto_extension_ieee754()
-{
-	return sqlite3_auto_extension(sqlite3_ieee_init);
-}
-
-int csqlite_sqlite3_cancel_auto_extension_ieee754()
-{
-	return sqlite3_cancel_auto_extension(sqlite3_ieee_init);
-}
-
-void sqlite3_series_init();
-int csqlite_sqlite3_auto_extension_series()
-{
-	return sqlite3_auto_extension(sqlite3_series_init);
-}
-
-int csqlite_sqlite3_cancel_auto_extension_series()
-{
-	return sqlite3_cancel_auto_extension(sqlite3_series_init);
-}
-
-void sqlite3_shathree_init();
-int csqlite_sqlite3_auto_extension_sha3()
-{
-	return sqlite3_auto_extension(sqlite3_shathree_init);
-}
-
-int csqlite_sqlite3_cancel_auto_extension_sha3()
-{
-	return sqlite3_cancel_auto_extension(sqlite3_shathree_init);
-}
-
-void sqlite3_uuid_init();
-int csqlite_sqlite3_auto_extension_uuid()
-{
-	return sqlite3_auto_extension(sqlite3_uuid_init);
-}
-
-int csqlite_sqlite3_cancel_auto_extension_uuid()
-{
-	return sqlite3_cancel_auto_extension(sqlite3_uuid_init);
-}
