+++ Sources/CSQLiteExtensions/decimal.c
