+++ Sources/CSQLiteExtensions/ieee754.c
