+++ Sources/CSQLiteExtensions/series.c
