+++ Sources/CSQLiteExtensions/uuid.c
